requirements-dev.txt
requests>=2.28
pytest>=7.0
python-dotenv>=1.0
tenacity>=8.2
pydub>=0.25
